requirements.txt
# Web Framework
fastapi
uvicorn[standard]
python-multipart

# ML & Vision (CPU-ONLY)
--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.2.1+cpu
torchvision==0.17.1+cpu
ultralytics
opencv-python-headless

openai
langchain
langchain-openai
langchain-community
langchain-core
langgraph
tiktoken

# Vector Search
faiss-cpu
rank-bm25

# PDF & Data
reportlab
pillow
pandas
numpy

# Utilities
python-dotenv
requests
pydantic
huggingface-hub
